First draft of poly node
Whew this was a tough one
After some troubleshooting, learned that voices can cancel themselves out so I need to put a slight offset on the duration. For now, I don't think I need to enforce this in code and it will come up again in my internal testing.
Need to reorder the inlets.

this will need refactoring...