Fix Metadata
I'm a little hazy on the MetasoundArrayNodesPrivate namespace, this comes from adapting the first template headers from the pattern in UE source which involved some heavy troubleshooting at the time. I should go back to those array nodes and check for redundancy / check how this was done in the MS source.

For now, this builds on Mac so let's try on Windows...